guiのリファクタ
