+++ src/gui/assignment_viewer.py
@@ -10,6 +10,8 @@ from api.model import Course
 from api.model import Attachment
 from api.model import SubmittedAttachment
 
+from gui.util import skip_request, get_year_and_semester
+
 
 class AssignmentCard(BaseModel):
     title: str
@@ -72,32 +74,6 @@ class AssignmentCard(BaseModel):
 
         st.markdown(card, unsafe_allow_html=True)
         return
-
-
-def get_year_and_semester() -> tuple[str, str]:
-    current_date = datetime.now()
-    year_int: int = current_date.year
-    if current_date.month < 4:
-        year_int -= 1
-    year: str = str(year_int)
-    if 4 <= current_date.month < 9:
-        semester = "前期"
-    else:
-        semester = "後期"
-    return year, semester
-
-
-def skip_request(course: Course, year: str, semester: str) -> bool:
-    if course.title == "Home":
-        print(f"{course.title}skipped")
-        return True
-    if course.title[1:5] != year:
-        print(f"{course.title}skipped")
-        return True
-    if course.title[5:7] != semester:
-        print(f"{course.title}skipped")
-        return True
-    return False
 
 
 def main() -> None:
